geoscience/image-impeccable/Igor Ivanov/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

numpy==1.26.3
scipy==1.12.0
pandas==2.2.0
pyarrow==17.0.0
scikit-learn==1.2.2
pillow==10.3.0
matplotlib==3.7.4
torch==2.1.2
torchvision==0.16.2
timm==0.9.2
segmentation_models_pytorch==0.3.3
scikit-image==0.24.0
